requirements.txt
langgraph
langchain-core
langchain-ollama
ollama
python-dotenv
